requirements.txt
numpy
fastapi[standard]
pydantic
uvicorn
google-api-python-client
python-dotenv
